KernelSU: Import SUSFS v1.4.2-lite * [1/2]
To be completely transparent in my opinion, SUSFS is one giant,
gnarly hack the likes of which I would generally not be interested
in including with this kernel, but these are trying times that we
live in and the ability to make KernelSU programmatically invisible
is undeniably useful.

That being said, the scope of SUSFS has ballooned over time to
encompass hiding all manner of mods that I have no interest in
using, or supporting.

For this reason, I've made a cursory pass through the SUSFS
codebase in order to restrict the amount of hackery going on in
this kernel. What remains is strictly what is required to hide
KernelSU, and try_umount because it integrates with KSU
functionality, although it can still be leveraged to hide any
desired path or mount point.

Any changes beyond that are purely superficial refactoring to
improve code clarity/readability, and correct typos.

[@0ctobot: Import based on:
https://gitlab.com/simonpunk/susfs4ksu/-/tree/gki-android14-5.15]

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>